+++ src/main/java/com/market/controller/dto/annotation/Password.java
@@ -1,4 +1,4 @@
-package com.market.global.annotation;
+package com.market.controller.dto.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,7 +6,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import com.market.global.annotation.validator.PasswordValidator;
+import com.market.controller.dto.annotation.validator.PasswordValidator;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
