+++ src/main/java/com/market/repository/mapper/RedisTemplateMapper.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Repository;
 
-import com.market.repository.interfaces.RedisRepository;
+import com.market.application.repository.interfaces.RedisRepository;
 
 import lombok.RequiredArgsConstructor;
